FIX: missing parameter for a debug print.  ENH: changed buffer.getDebugSize() return type to size_t.  TESTING: changed to use string.append instead of stringstream to test if that is the cause of keeneland segv's.

+++ src/runtime/DataBuffer.h
@@ -31,7 +31,7 @@ public:
 	DataBuffer(int _capacity, cciutils::SCIOLogSession *_logsession = NULL);
 
 	// for data addition
-	virtual int debugBufferSize() { return buffer.size(); };
+	virtual size_t debugBufferSize() { return buffer.size(); };
 
 //	bool isEmpty() { return getBufferSize() == 0; };
 	virtual bool isFull() { return buffer.size() >= capacity; };
